se realizo todas las user stories programables

+++ src/main/java/pe/edu/upc/smartharvest/repositories/IUsersRepository.java
@@ -3,7 +3,9 @@ package pe.edu.upc.smartharvest.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pe.edu.upc.smartharvest.entities.Users;
+
 @Repository
-public interface IUsersRepository extends JpaRepository<Users, Integer> {
+
+public interface IUsersRepository extends JpaRepository<Users, Long> {
     public Users findOneByUsername(String username);
 }
